requirements.txt
djangorestframework
psycopg2-binary
python-decouple
django-filter